Always update the progress bars even when stats are unavailable.
In case of an error (e.g. a device gets removed or some
misdetection by the kernel), stats may become unavailable,
but the progress bars still show the old values, and the
plugin may seem frozen. Because we don't like it to appear
frozen, always update the bars and set their values to 0
when no stats are available.

Signed-off-by: Landry Breuil <landry@xfce.org>